% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qfunction.R
\name{Q1}
\alias{Q1}
\title{get maximum of V2}
\usage{
Q1(outcome, tx1, tx2, cov1, cov2)
}
\arguments{
\item{outcome}{is CD4 # at the end}

\item{tx1,tx2}{are treatment choices at each decision point}

\item{cov1,cov2}{are two covariate of each individual}
}
\value{
maximum value of outcome
}
\description{
This step is try to get back the trial history
}

