\name{qfunction-package}
\alias{qfunction-package}
\alias{qfunction}
\docType{package}
\title{
Q function
}
\description{
 Regressing outcome at time 2 on coveriates and previous treatment choice, in order to get the optimal 
 treatment at time 2 which can maximize outcome. Then go back to do the same to time 1.

}
\details{
\tabular{ll}{
Package: \tab qfunction\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-11\cr
License: \tab Yuting,Biostat,Brown\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Yuting Zhang

Maintainer: Yuting Zhang<yuitng_zhang@brown.edu>

}
\references{

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{

}
\examples{

}
